Create tie breaker.java
Tie Breaker
Given an array of names arr[] of candidates in an election. A candidate name in array represents a vote casted to the candidate. You need to find the candidate with maximum votes recieved in the election. It may happen that two or more candidates may have same votes, in that case you need to return the name of the candidate which comes first lexicographically in dictionary.