Create frequency count.java
Frequency Count

Given an integer array arr[], you need to find frequency of each element and return in an arraylist in the order they appear.

Examples:
Input: arr[] = [10, 5, 20, 5, 10, 5]
Output: 2 3 1
Explanation: The freq are as follows: 10:2, 5:3, 20:1, so we print the freq in the order of appearance of the elements. We print freq of any element only once.